public override void Serialize<TBufferWriter>(ref MemoryPackWriter<TBufferWriter> writer, scoped ref G.IEnumerable<T>? value)=>writer.Write(Formatter,value); テスト用のSet<Key,Value>の型のシリアライズ設定が間違えていた。 <object>ではなく<T>としてシリアライズするようにした。 TypeString,StringTypeの短縮処理が間違っていたので直した。 シリアライズNewArrayのnullをカバー Utf8JsonではResolver.GetFormatterでStandardResolver.Instance.GetFormatterが使えた。 Serializableクラスが現れた場合DynamicObjectResolver.GetFormatterを使うことでDisplayClassのthisフィールドを避けてシリアライズできるようになった。 MemoryPackでExpression<T>ではなくLambdaExpressionをシリアライズしようとしたらExpression<T>に登録されていたのでGetRegisteredFormatter処理を変えた 変換_局所Parameterの先行評価にBlock処理が入っていなくて入れてListスコープParameterを設定した。 IEnumerable<int>,IEnumerable<double>と複数継承した場合、かつ元クラスが自作じゃなかった(<UnionByIterator>d__300)場合シリアライズできない。

+++ LinqDB/Sets/ImmutableSet`1.cs
@@ -208,7 +208,7 @@ public abstract class ImmutableSet<T>:ImmutableSet, IEnumerable<T>,IEquatable<IE
     /// <returns>追加に成功すればtrue、失敗すればfalse。</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal virtual bool InternalAdd(T Item) {
-        var HashCode = typeof(T).IsNullable()?(long)(uint)Item!.GetHashCode():Item is not null?(long)(uint)Item.GetHashCode():0;
+        var HashCode = typeof(T).IsNullable()?(uint)Item!.GetHashCode():Item is not null?(long)(uint)Item.GetHashCode():0;
         if(this.InternalAdd前半(out var 下限,out var 上限,out var TreeNode,HashCode)) {
             var Comparer = this.Comparer;
             LinkedNodeT LinkedNode = TreeNode;
